Fixed integration between Applicaion.py and Entry.py
I commented out the integration of the splash screen on Application.py due to it causing crashes. We'll need to fix that.

Expanded Entry screen to have entries for first name, last name, and code name. We should pull the id and key from the counter variable (i) when we are ready to use that data.